\name{train_linear_model}
\alias{train_linear_model}
\title{train_linear_model}
\usage{ model <- train_linear_model(data)
}
\description{
Inside the function we create a formula that describes the relationship between the variables in our dataset and the outcome we're trying to predict. This formula tells the model how to use the data to make predictions.
}
\examples{
model <- train_linear_model(data)
}
