\name{visualize_data}
\alias{visualize_data}
\title{visualize_data}
\usage{visualize_data(data, synthetic_data)}
\description{provides a visual comparison between original and synthetic datasets.    This function orchestrates the creation of histograms to depict the distributions of each variable in both datasets and scatterplot matrices to visualize pairwise relationships.
}
\examples{
visualize_data(data, synthetic_data)
}
